backend/requirements.txt
-i https://pypi.org/simple
amqp==2.6.0
asgiref==3.2.10
babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
celery==4.4.2
certifi==2020.6.20
chardet==3.0.4
coreapi==1.32.3
coverage==5.2
decorator==4.4.2
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cors-headers==3.4.0
django-environ==0.4.5
django-extensions==2.0.7
django-filter==2.3.0
django-guardian==2.3.0
django-livereload-server==0.3.2
django-phonenumber-field[phonenumbers]==4.0.0
django-rest-auth==0.9.5
django-timezone-field==4.0
django==3.0.7
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djangorestframework==3.11.0
factory-boy==2.12.0
faker==4.1.1
gunicorn==20.0.4
idna==2.10
ipython-genutils==0.2.0
ipython==7.16.1
itypes==1.2.0
jedi==0.17.1
kombu==4.6.11
markdown==3.0.1
mysqlclient==1.4.6
parso==0.7.0
pexpect==4.8.0; sys_platform != 'win32'
phonenumbers==8.12.6
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pygments==2.6.1
pyjwt==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.1
pytz==2018.5
pyyaml==5.3.1
redis==3.5.3
requests==2.24.0
six==1.11.0
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
tornado==6.0.4
traitlets==4.3.3
typing==3.6.4
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
wcwidth==0.2.5
